+++ pkg/ffuf/request.go
@@ -3,6 +3,7 @@ package ffuf
 // Request holds the meaningful data that is passed for runner for making the query
 type Request struct {
 	Method   string
+	Host     string
 	Url      string
 	Headers  map[string]string
 	Data     []byte