+++ CHANGELOG.md
@@ -4,9 +4,17 @@
   - New
     - New CLI flag `-maxtime-job` to set max. execution time per job.
     - Changed behaviour of `-maxtime`, can now be used for entire process.
+    - A new flag `-ignore-body` so ffuf does not fetch the response content. Default value=false.
+    - Added the wordlists to the header information.
+    - Added support to output "all" formats (specify the path/filename sans file extension and ffuf will add the appropriate suffix for the filetype)
+
   - Changed
     - Fixed a bug related to the autocalibration feature making the random seed initialization also to take place before autocalibration needs it.
-    
+    - Added tls renegotiation flag to fix #193 in http.Client
+    - Fixed HTML report to display select/combo-box for rows per page (and increased default from 10 to 250 rows).
+    - Added Host information to JSON output file
+    - Fixed request method when supplying request file
+
 - v1.0.2
   - Changed
     - Write POST request data properly to file when ran with `-od`.